Objet Context

🌐
Context object

Les objets context sont utilisés par luz pour l'échange d'information entre méthodes, métriques et callbaks.

Détails

L'objet ctx est utilisé par luz pour partager des méthodes entre la boucle d'entraînement, les callbacks, les méthodes du modèles et les métriques. La table ci-dessous liste les informations disponibles dans ctx par défaut. Les autres callbacks peuvent potentiellement modifier ces attributs et en ajouter d'autres.

Attribute	Description

verbose	La valeur (TRUE ou FALSE) affectée à l'argument verbose dans fit .

accelerator	L'objet Accelerator utilisé pour distribuer la charge de calcul aux cartes accélératrices, les données, etc. It assumes the value passed to the accelerator parameter in fit.

model	Un objet nn_module initialisé qui sera entraîné pendant le process de fit.

optimizers	Liste nommée d'optimateurs utilisés pendant l'entraînement.

data	Le dataloader en cours d'usage. Pendant l'entraînement, c'est ctx$train_data, pendant l'évaluation, c'est ctx$valid_data. It can also be the prediction dataset when in predict.

train_data	Chargeur de données fourni comme argument data à fit. Modifié pour fournir les données au device cible.

valid_data	Chargeur de données fourni comme argument valid_data à fit. Modifié pour fournir les données au device cible.

min_epochs	Nombre minimum d'époques que le modèle sera entraîné.

max_epochs	Nombre maximum d'époques que le modèle sera entraîné.

epoch	Époque actuelle de l'entraînement.

iter	Itération actuelle de l'entraînement. Il est réinitialisé à chaque époque et lorsqu'on passe d'une entraînement à une validation.

training	Si le modèle est en mode entraînement ou validation. Voir également help("luz_callback_train_valid")

callbacks	Liste des callbacks qui seront appelés pendant la procédure d'entraînement. C'est l'union de la liste passée à l'argument callbacks et des callbacks par défaut.

step	la closure qui sera utilisé pour faire un step du modèle. Elle est utilisée pour l'entraînement et la validation. Ne prends pas d'argument mais a accès à l'objet ctx.

call_callbacks	Appel des callbacks par leur nom. Par exemple call_callbacks("on_train_begin") appelle tous les callbacks qui fournissent des methodes pour ce point.

batch	Dernier lot obtenu par le chargeur de données. le lot est list() de 2 éléments, l'un est utilisé comme input et l'autre come target.

input	Premier élément de la liste du dernier lot obtenu par le chargeur de données actuel.

target	Second élément de la liste du dernier lot obtenu par le chargeur de données actuel.

pred	Dernières prédictions obtenues par ctx$model$forward . Note: cela peut être potentiellement modifié par les appels précédemment exécutés. Ne sera pas disponible si vous utilisez un pas d'entraînement personnalisé.

loss_fn	La fonction de coût active, qui sera minimisée pendant l'entraînement.

loss	La dernière valeur mesurée de la fonction de coût Note: Ne sera pas disponible avec un pas d'entraînement personnalisé.

opt	L'optimiseur en cours, c.a.d. celui qui sera utilisé pour le prochain step pour la mise à jour des paramêtres du modèle.

opt_nm	Le nom de l'optimiseur en cours. Par default, c'est opt, sauf si votre modèle utilise plusieurs optimiseurs suivant le jeu de paramêtres à optimiser.

metrics	Une list() des objet de métrique en cours updateés à chaqe on_train_batch_end() ou on_valid_batch_end(). Voir aussi help("luz_callback_metrics")

records	Une list() enregitrant toutes les valeurs des métriques d'entraînement et de validation. Voir aussi help("luz_callback_metrics"). Contient aussi les métrique sde profiling. Voir help("luz_callback_profile") pour plus d'information.

handlers	Une list() nommée des handlers transférée à rlang::with_handlers() pendant la boucle d'entraînement, qui permet de gérer des erreurs ou des conditions levées par d'autres callbacks.

epoch_handlers	Une list() nommée des handlers utilisée avec rlang::with_handlers(). Ces handlers sont utilisés à l'intérieur de la boucle de l'époque, et vous permettent de gérer des évènements spécifiques à l'époque, sans pour autant arrêter l'entraînement.

Attributs du contexte

Voir également