Callback de reprise de l'entraînement du modèle.

🌐
Resume training callback

Ce callback vous permet de reprendre l'entraînement du modèle.

Utilisation

luz_callback_auto_resume(path = "./state.pt")

Arguments

path

Chemin où sauvegarder les fichiers d'instantané du modèle.

Détails

Quand on l'utilise, les poids du modèle et l'état de l'optimiseur sont sérialisés à la fin de chaque époque. Si quelque chose échoue pendant l'entraînement, réexécuter le même script reprendra l'entraînement du modèle à l'époque qui suit la dernière époque sérialisée.

Note

En général, vous voudrez ajouter ce callback en dernier dans la liste de callbacks, car ainsi, l'état sérialisé est susceptible de contenir toutes les modifications possibles apportées par d'autres callbacks à 'on_epoch_end'. Le weight par défaut de ce callback est Inf.

Voir la vignette sur les instantanés pour plus de détails.

Customizing serialization

Par défaut, le modèle, l'état de l'optimiseur et l'état des callbacks sont sérialisés. Les callbacks peuvent être utilisés pour personnaliser la sérialisation en implémentant les méthodes state_dict() et load_state_dict(). Si ces méthodes sont implémentées, alors state_dict() est appelée à la fin de chaque époque et load_state_dict() est appelée lorsque le modèle reprend.

Exemples