Sauvegarde un instantanné du modèle

🌐
Checkpoints model weights

Utilisation

luz_callback_model_checkpoint(
  path,
  monitor = "valid_loss",
  save_best_only = FALSE,
  mode = "min",
  min_delta = 0
)

Arguments

path

Chemin pour sauvegarder le modèle sur disque dur. Le chemin est interpolé avec glue, donc vous pouvez utiliser tout attribut dans le ctx en utilisant '{ctx$epoch}'. Par exemple, les valeurs epoch et monitor sont déjà présentes dans l'environnement. Si le chemin spécifié est un répertoire (se termine par / ou }), alors les modèles sont sauvegardés avec le nom fourni par \verb{epoch-{epoch:02d}-{self$monitor}-{monitor:.3f}.pt}. Voir les exemples. Vous pouvez utiliser \code{\link[=sprintf]{sprintf()}} pour formater rapidement les valeurs, par exemple:\code{'{epoch:02d}'}. monitorUne chaîne au format <set>_<metric> où <set> peut être 'train' ou 'valid' et <metric> est l'abréviation de toute métrique suivie pendant l'entraînement. Le nom de la métrique est insensible à la casse. save_best_onlySi TRUE, les modèles ne sont sauvegardés que si ils améliorent un modèle précédemment enregistré. modeSpécifie la direction considérée comme une amélioration. Par défaut, 'min' est utilisé. Cela peut également être 'max' (lorque on cherche à maximiser la métrique) et 'zero' (plus proche de zéro est mieux). min_deltaDifférence minimale à considérer comme une amélioration. Seulement utilisée lorsque save_best_only=TRUE.

Sauvegarde l'instantanné du modèle selon la métrique spécifiée et le comportement. mode et min_delta ne sont utilisés que lorsque save_best_only=TRUE. save_best_only écrasera les modèles enregistrés si le paramètre path ne différencie pas les noms des modèles par epochs.Voir la vignette sur les instantanés pour plus de détails. luz_callback_model_checkpoint(path= "path/to/dir") luz_callback_model_checkpoint(path= "path/to/dir/epoch-epoch:02d/model.pt") luz_callback_model_checkpoint(path= "path/to/dir/epoch-epoch:02d/model-monitor:.2f.pt") Autres callbacks luz: luz_callback_auto_resume(), luz_callback_csv_logger(), luz_callback_early_stopping(), luz_callback_interrupt(), luz_callback_keep_best_model(), luz_callback_lr_scheduler(), luz_callback_metrics(), luz_callback_mixup(), luz_callback_model_checkpoint(), luz_callback_profile(), luz_callback_progress(), luz_callback_resume_from_checkpoint(), luz_callback_train_valid(), luz_callback() luz_callbacks