Mixup logic

Logic underlying luz_callback_mixup().

Utilisation

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Valeur de retour

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Détails

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

Voir également

Exemples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-1.0219  1.0113 -0.0940 -1.6500 -0.3305 -0.6906 -1.2601 -0.6884  0.2464 -0.0005
#>  0.5471 -2.4207  0.5743 -0.2266 -0.6439  0.0932 -1.9379  0.7853  0.2224 -0.9831
#>  0.3248  0.2434 -0.9238 -0.1459 -0.9081  1.3515  1.1218 -0.8036 -0.8745 -0.8808
#> -0.5148  0.5294  0.9522 -1.8993 -0.7130  1.3281  0.2129 -2.2482 -0.9391  0.3687
#>  1.2066 -0.9892  0.0343 -1.4138  0.5774 -1.0285  0.0542 -0.9878 -0.4731 -0.0300
#>  1.3105  0.2263  0.6712  1.3361  0.2601  0.6078 -1.5534  0.7247  0.9028  0.0986
#> -2.2320 -0.0238  1.0189 -2.0779 -0.1659 -1.3138  0.0262 -0.1378  0.3171 -1.5599
#> -0.6446 -1.1136  0.0303  0.8125  1.1367 -1.3162  0.7353 -0.0552  1.0776 -0.6043
#> -1.2999 -1.0949 -0.9401  0.1227  0.3505 -1.6980  1.4118  1.5901 -0.5988 -0.0868
#> -1.4921  1.2557 -0.7430 -0.8483 -0.3979 -0.3705 -0.2268  0.4707 -0.7605  1.1547
#> 
#> Columns 11 to 20-1.1668  0.1162  0.9077  0.7312 -0.2723  0.4069  1.3628 -0.1470  0.5495  0.0098
#> -0.6975  1.0777  0.4778  0.9463  0.3886  0.4671  0.3516 -0.3060 -1.1069 -0.7137
#>  0.0178  0.0921  0.0561  1.2716  0.4801 -0.3385  1.3283 -0.1785  0.3829  0.7733
#> -0.7844 -0.9703  0.5923  0.2855  1.0619 -0.6694  1.4970 -0.7494  0.2365  1.8229
#> -0.8547  1.2066 -1.1198 -0.5404  0.7490 -0.7207 -0.6811 -0.3703 -0.3613  0.0189
#> -1.2557  1.8966 -1.8813 -0.3429 -0.1494  0.4357 -0.2629 -1.4305  0.6112  1.4137
#>  1.3427 -2.3419  0.4976  0.2220  1.9451 -1.0152  1.7208 -0.6438  1.5080  0.8746
#>  0.9930  0.1296 -3.0018  0.1763  0.8293  0.8536 -1.0155  0.7166 -0.2401 -0.9495
#>  1.0368 -0.8900 -2.4233 -2.2316 -0.3100  0.5117  0.6931 -1.3126  0.7297 -0.2550
#> -0.3536 -1.3218  1.8145  0.1789  0.6922  0.1205  0.1473  0.5266  2.1056 -1.1100
#> 
#> Columns 21 to 30 1.7423  0.2632  0.5591  0.0998 -0.2854 -1.6903  0.1103 -0.1822  1.2373 -0.6043
#>  0.6074  0.6474 -0.7372  0.3240 -1.0357 -0.6335  0.4153 -0.4554  1.1547 -0.5202
#>  0.6920 -0.6583 -0.2681  0.9511  1.4217  0.5449 -0.0426  0.7534  0.7748 -1.5187
#>  1.2357 -0.1587 -0.2108  0.7269  0.0758  0.9779  1.1043 -0.4395 -0.5551  0.0590
#>  0.6855 -0.0015  1.1463 -1.0780 -0.2615  0.1997 -0.5259 -0.6687 -0.6165 -0.1908
#> -0.4808  0.3812 -0.9153 -0.9570 -0.1156 -0.3548  0.4312 -0.5344  1.2807  0.2402
#> -0.0849  0.3703 -0.4683  0.4137 -0.0405  0.5717 -1.1913 -0.0671  0.1973 -0.5112
#>  0.1207 -0.4110  0.7031  0.3742  0.9250 -0.8311  0.5832  0.6633 -0.5984  1.1426
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.5744
#>  0.8664
#> -0.5058
#>  0.2268
#> -0.3215
#>  1.3882
#>  0.2293
#> -0.9115
#>  0.6881
#>  0.6903
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.6903
#> -0.3215
#>  0.2268
#> -0.5058
#>  1.3882
#>  0.8664
#>  0.2293
#>  0.6881
#> -0.9115
#> -1.5744
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>