Mixup logic

Logic underlying luz_callback_mixup().

Utilisation

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Valeur de retour

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Détails

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

Voir également

Exemples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-1.2872e+00  1.4136e+00 -1.8332e+00 -3.3090e-01 -6.2604e-02  4.9432e-01
#> -1.3356e+00  1.0655e+00 -1.7194e+00  6.3000e-01 -6.8845e-01 -5.4155e-01
#>  2.6973e-01 -7.7269e-01  6.5100e-02  1.7287e-01  3.1672e-01 -2.2210e-01
#> -2.2688e+00 -6.1553e-02  6.5188e-01 -1.0883e+00  6.5451e-01 -6.8778e-01
#> -1.2350e-01  2.0100e+00 -1.8573e+00 -1.6985e+00 -9.1957e-01 -1.3777e-01
#> -2.0071e+00  1.5060e+00 -5.5942e-01  6.6334e-01 -6.9282e-01 -5.1689e-01
#>  6.3580e-01  2.0953e+00 -5.1135e-01  1.2231e+00  5.8742e-01 -6.1801e-01
#>  3.5529e-01  9.4825e-02  9.1851e-02  1.2833e-01  4.1424e-01 -1.0497e+00
#>  1.5839e-01  8.1604e-02  2.5780e-01  3.3110e-01 -2.8754e-01  7.9522e-01
#> -1.4099e+00  6.3899e-01  1.4118e+00 -1.1758e-02  6.3748e-01  6.4040e-01
#> 
#> Columns 7 to 12 3.8770e-01 -2.6008e-01  6.1713e-01  9.0991e-01  2.8134e-01 -1.6521e-01
#> -4.2641e-01 -2.4860e-01 -1.0575e+00 -8.6838e-02  1.2945e+00 -2.4839e-01
#> -1.6721e-01 -1.6053e+00 -2.7452e-01  4.3914e-01  2.1941e-01 -6.5136e-01
#>  1.3890e-01 -1.0123e+00 -9.5899e-01 -2.0832e-02  5.2870e-01 -9.1342e-01
#>  3.2569e-01 -1.1673e+00  4.4014e-01  6.5711e-02  4.7709e-01  9.2014e-01
#>  8.8352e-01 -7.7464e-01 -4.7670e-01  2.2341e-01 -1.0183e+00 -9.8432e-02
#>  6.0511e-01 -1.6345e+00  1.3650e+00 -5.8359e-03  5.9221e-01 -8.7911e-02
#>  4.5094e-01 -2.7169e-01  1.5902e-01  7.3119e-01 -4.4887e-01 -1.5678e+00
#>  4.3740e-01 -1.5929e+00 -1.1334e+00 -2.8646e-01  1.6739e-02 -6.5084e-01
#> -1.0448e-01  1.4012e+00  1.2077e+00 -1.1542e+00 -4.9482e-01  3.2530e-01
#> 
#> Columns 13 to 18-6.0305e-01 -1.6426e+00 -1.0482e+00  5.6974e-01 -5.8220e-01  2.0411e-01
#> -8.3677e-01  4.3185e-01  9.4704e-01 -5.2285e-01 -8.1141e-01  4.4026e-01
#> -1.6233e+00 -6.1969e-01  7.2057e-01 -1.3143e-01  1.0091e+00  2.2572e+00
#> -3.9286e-01  2.0943e+00  5.6198e-01  8.2051e-03 -8.4941e-01 -1.2467e+00
#> -5.7834e-01 -1.1068e+00 -4.8212e-02  5.2835e-02 -3.9035e-01 -1.9520e-01
#> -1.2165e+00 -7.3154e-01  8.7099e-01  1.2784e+00 -6.2991e-02 -2.2427e-02
#> -1.7835e-01  3.0908e-01  6.6114e-01 -1.6927e-01  8.8997e-01 -1.4978e+00
#>  6.2390e-01 -6.4348e-01  7.5932e-01 -8.7637e-01  2.5107e+00  3.3675e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.4264
#>  0.2039
#> -0.1360
#>  0.3177
#>  1.1650
#> -0.3436
#> -0.9074
#>  0.0631
#> -0.4187
#>  0.3716
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.1650
#> -0.3436
#> -0.4187
#>  0.3716
#>  0.2039
#>  1.4264
#>  0.0631
#> -0.9074
#>  0.3177
#> -0.1360
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>