Mixup logic

Logic underlying luz_callback_mixup().

Utilisation

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Valeur de retour

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Détails

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

Voir également

Exemples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-2.0400 -0.4050  0.1559  2.4608  0.2493 -0.3130  0.8957  0.5735  1.1698 -0.3578
#>  1.1304 -0.0944 -0.1752  0.2510  0.1319  0.2993 -0.1820  0.4057  1.2221  0.7015
#>  0.5451  0.0678 -0.5012 -0.4074  0.5176  1.3486  0.1000 -1.5500 -1.1298  0.4487
#> -1.5556  0.8506 -0.1222  0.9824 -1.1029  0.2395  0.2003  1.4042  1.4828  1.2765
#>  0.9618 -0.9860  0.9524 -0.6092 -0.9836  0.5278 -0.2089  0.3494 -0.6282 -0.1344
#> -1.3303 -0.4453 -0.4870  0.6301 -0.2509  0.3089  0.6395  0.4004  0.9290  0.0091
#>  0.2330  0.1528  0.5994 -1.5078  1.5950  0.6023 -0.8326 -0.0861  1.0625 -0.5327
#>  0.5917 -0.3582  1.2817  0.6379  1.2267  1.5551  0.1912  0.4282  1.5047  0.4810
#>  0.8246 -0.9920 -0.5457  1.2660 -1.6677  0.1742  0.0690  0.6113 -0.6581  0.7839
#>  0.1424  0.1216  0.3662  0.6478 -1.4853  1.0414 -0.3464  0.4453  0.3494  0.4740
#> 
#> Columns 11 to 20-2.2936 -0.1824  2.0569 -0.3820 -0.9996  0.3886  1.1339 -0.2380  3.0547  0.8269
#>  0.4500 -0.3718 -1.1122 -1.0997  1.9887 -0.3030  0.2904 -0.9081 -0.0306  0.0739
#> -0.2058 -1.4290  1.0899 -0.6090  0.9888  0.6954  0.6416  0.0791  1.4157  0.4972
#>  1.0219 -0.5373  0.3910 -0.5250 -0.7731 -0.2956  0.0886 -0.4116  1.1038 -0.6246
#>  0.0871 -0.0480 -0.2096 -0.0709  0.0051  0.6967 -0.9611 -0.5374  0.8968 -0.3818
#>  0.0048 -0.5442  0.7165 -2.1656 -0.5144 -0.6081 -0.5020 -0.3860 -0.5855  0.5356
#>  0.6541  0.8567 -0.3486 -0.6749  0.7755  1.1943  1.1379  0.1117  0.9293 -0.4729
#> -1.3280  0.3017  2.3455  0.5280  0.5971 -0.5272  1.6935  0.2778  0.6237  1.1485
#>  0.1914  1.6040 -1.6384  1.5908 -1.1638  0.0948  0.7864  0.5075  0.2942  0.6857
#>  0.4748 -0.1644 -1.1874  0.4511  1.1002  0.0803 -0.8023 -0.2995  1.5363  0.8152
#> 
#> Columns 21 to 30 0.3091 -0.0941  0.3104 -0.0051 -0.3104  0.9846  1.9519  0.1296  1.2590 -0.6406
#>  1.2387 -0.8399 -2.8053 -0.6213  1.5421 -0.6922 -0.5838 -1.1527 -1.1552 -1.6591
#>  1.0948 -0.2282  0.7736  1.2295 -0.2250  0.8985 -0.6488 -1.5692  0.2111 -0.7557
#> -0.5728 -0.1316  0.2413 -0.4548 -1.9155 -1.2015 -0.0433  1.6626 -0.1685  0.0845
#> -0.5769  2.5119 -0.5980  0.3207  0.2261  0.0443 -0.0170 -0.5767 -1.1563  1.0090
#> -0.8782  0.1986  0.0433  0.4608 -0.9568  1.0405  0.0187 -0.1325 -0.8270  0.4679
#> -0.6599 -1.1021 -0.9752  1.6342  0.1393  0.5507  1.2581  1.1209  0.5833  1.6129
#> -0.0312 -0.9114 -0.5325  1.7125 -0.1344  2.4477  0.4456  0.6251  1.1020  1.0448
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.4840
#> -0.0754
#> -0.1310
#> -0.5951
#> -1.2228
#>  0.7054
#>  0.3350
#> -3.4971
#>  0.5670
#> -0.4027
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2228
#>  0.7054
#> -0.0754
#>  0.5670
#>  0.4840
#> -0.5951
#> -3.4971
#>  0.3350
#> -0.4027
#> -0.1310
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>