Mixup logic

Logic underlying luz_callback_mixup().

Utilisation

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Valeur de retour

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Détails

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

Voir également

Exemples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-1.5268e+00  5.1536e-01  7.4108e-02  6.4892e-01  8.6822e-02 -1.2684e+00
#> -5.3850e-01 -6.5634e-01  9.0361e-01 -5.1436e-01  9.4255e-01 -6.8146e-01
#> -9.6889e-02 -1.6453e+00 -4.0988e-02 -4.4749e-02 -1.7255e+00  5.3909e-01
#> -1.2286e+00  6.8626e-01  5.4588e-01  7.8267e-02  2.0602e+00 -1.6268e-01
#> -6.2272e-01 -9.9191e-01  6.2950e-01  1.1859e+00  2.1013e+00  4.4429e-01
#>  7.8929e-01  1.7611e+00 -3.6892e-01  7.4986e-01 -7.7185e-01  9.9206e-01
#> -8.8559e-01  1.3609e+00 -9.8363e-01 -1.7787e+00  1.4882e+00 -1.0933e+00
#> -2.5812e+00  1.0455e+00 -1.5330e-01 -3.7939e-02 -3.1127e-01 -1.1030e+00
#> -9.1639e-01  8.1245e-01 -8.7876e-02 -5.6401e-01  2.9487e-01  3.8510e-01
#> -5.7782e-01 -1.2729e+00 -1.3130e-01 -2.4300e-01 -7.8326e-01  9.9772e-01
#> 
#> Columns 7 to 12 8.5506e-01 -6.2027e-01 -2.6000e-01  9.8775e-01 -4.1108e-02  1.4002e+00
#> -4.9878e-02  3.0741e-03 -1.1857e+00 -1.4630e+00 -1.9197e+00  1.2376e-01
#> -5.0978e-01  3.5396e-01 -8.3739e-02 -4.9253e-01  4.6131e-01  5.8082e-01
#> -5.5011e-02  2.0574e-01  8.0225e-01 -3.0343e-01 -9.7393e-01  9.7797e-02
#>  1.3968e+00 -8.5462e-01 -2.5913e-01 -1.1038e+00 -6.0814e-01 -2.3507e-01
#>  7.1143e-02 -1.3882e+00 -1.0824e+00  6.8581e-01 -1.8652e-01 -5.0299e-02
#>  5.2664e-02  7.8349e-02  1.0790e+00 -1.5425e+00 -4.7431e-01  1.2104e+00
#>  2.3014e-01 -1.1571e+00 -1.6657e-01  6.3749e-01 -9.8436e-01  4.6521e-01
#>  4.5259e-01 -4.4853e-02  9.2211e-01 -2.1189e-01 -2.3340e-01  1.2505e+00
#> -9.1032e-01 -6.9542e-02 -9.6187e-01 -8.4561e-02 -6.4491e-01 -1.3589e+00
#> 
#> Columns 13 to 18-4.7482e-01  4.7401e-01  1.4082e+00  8.5213e-04 -1.6174e+00 -3.1047e-01
#> -7.8454e-01 -8.6231e-01 -5.5410e-01 -3.8356e-01 -2.0011e+00 -3.7473e-01
#> -1.3570e+00  8.7403e-01  1.0426e+00  2.0473e+00 -2.0279e-01 -1.4119e+00
#>  5.9445e-01  6.1242e-01 -6.4913e-01  9.5313e-01 -6.6291e-01 -1.1395e+00
#> -3.4558e-01  5.6593e-01 -6.2317e-01 -2.7217e-01 -8.6587e-01  5.6554e-01
#>  4.3890e-01 -6.9358e-01 -6.0313e-01  9.9183e-02  7.3615e-01 -8.4522e-01
#>  6.2137e-02 -9.0796e-01  2.3196e-01 -1.3792e+00 -3.1700e-03  3.6899e-01
#>  1.0794e+00 -6.2193e-01  1.7975e+00  8.4156e-01 -5.8841e-03  6.3232e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -3.0369
#>  0.3215
#> -0.7482
#> -0.3114
#> -1.8147
#> -1.2726
#> -0.5514
#> -0.1166
#> -1.7876
#>  0.5690
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.3215
#> -1.2726
#>  0.5690
#> -0.3114
#> -1.7876
#> -1.8147
#> -3.0369
#> -0.1166
#> -0.5514
#> -0.7482
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>